WIP correctly switch modes on reset and start
There needs to be a 100% reliable way to determine which window is the
producer window and which is the prompt. Some ideas:

- Use window focus
- Use localstorage update lag
- An explicit button
- Listen on a broadcast channel

One big problem currently: when you press start from one window, it
ought to immediatly update its state to "producer mode", but in fact the
timer emits a few times before the state update carries through. This
results in the first few componentDidUpdate calls having no way to
differentiate between the active window and the inactive window.