Fix false starts after reset
Since the reset flag was only reset after the timer started,
componentDidUpdate was being called with the correct conditions to call
uninitializedMode() after a new recording was started, causing
producerMode to never appear.

This adds a new redux action to reset the reset flag. We call this
action as soon as we revert to uninitializedMode, so the reset flag is
only ever flipped to true for a moment.