Don't set state to uninitialized on start
On a first-time use of the app, when there's no existing Redux state,
wasReset is set to an empty array. This evaluates to true in the
previous conditional, resulting in a false start for new users of the
app. We're actually looking for an explicit true value.

Also adds a missing semicolon.